Reorganize folder (still issues with imports when from entrypoint)
