+++ lib/entrypoint.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 if [ "${TEST_CASE_RUN}" == "true" ]; then
-  python test.py
+  python ./superlinter/test.py
 else
-  python SuperLinter.py --cli
+  python ./superlinter/run.py
 fi
