+++ lib/superlinter/LinterTemplate.py
