+++ lib/superlinter/__init__.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python3
+from .SuperLinter import SuperLinter
+from .LinterTemplate import LinterTemplate
+from .linters import *
+
+__all__ = [
+    'SuperLinter',
+    'LinterTemplate',
+]
