+++ lib/superlinter/run.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python3
+"""
+Run super-linter
+
+@author: Nicolas Vuillamy
+"""
+
+from lib.superlinter import SuperLinter
+
+# Run Super-Linter
+SuperLinter({'cli': True}).run()
