+++ lib/superlinter/tests/test.py
@@ -1,0 +1,13 @@
+#!/usr/bin/env python3
+"""
+Unit tests for Super-Linter
+
+@author: Nicolas Vuillamy
+"""
+import unittest
+
+from lib.superlinter.tests.test_superlinter.test_suite import suite
+
+if __name__ == '__main__':
+    # noinspection PyTypeChecker
+    unittest.TextTestRunner(verbosity=2).run(suite())
